requirements.txt
python-telegram-bot==12.7
lxml>=5.0.0
requests>=2.32.0
python-dotenv>=1.0.1
reportlab>=4.0.0
